backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
beanie==2.0.1
certifi==2026.1.4
click==8.3.1
fastapi==0.128.0
h11==0.16.0
idna==3.11
motor==3.7.1
pymongo==4.16.0
python-dotenv==1.2.1
starlette==0.50.0
typing_extensions==4.15.0
uvicorn==0.40.0
